// src/api/stetsom/model/publicCategory.ts
/**
 * Generated by orval v8.10.0 🍺
 * Do not edit manually.
 * Stetsom API
 * API documentation for Stetsom
 * OpenAPI spec version: 1.0.0
 */
import type { PublicLine } from "./publicLine";

export interface PublicCategory {
  id: string;
  name: string;
  slug: string;
  icon_library_id?: string;
  /** @nullable */
  icon_url?: string | null;
  lines: PublicLine[];
  /**
   * @minimum -9007199254740991
   * @maximum 9007199254740991
   */
  order: number;
  created_at: string;
  updated_at: string;
}